Move fee manager into home mediator (#462)
