+++ contracts/upgradeable_contracts/BasicAMBMediator.sol
@@ -15,6 +15,15 @@ contract BasicAMBMediator is Ownable {
     bytes32 internal constant REQUEST_GAS_LIMIT = 0x2dfd6c9f781bb6bbb5369c114e949b69ebb440ef3d4dd6b2836225eb1dc3a2be; // keccak256(abi.encodePacked("requestGasLimit"))
 
     /**
+    * @dev Throws if caller on the other side if not an associated mediator.
+    */
+    modifier onlyMediator {
+        require(msg.sender == address(bridgeContract()));
+        require(messageSender() == mediatorContractOnOtherSide());
+        _;
+    }
+
+    /**
     * @dev Sets the AMB bridge contract address. Only the owner can call this method.
     * @param _bridgeContract the address of the bridge contract.
     */
