+++ test/multi_amb_erc20_to_erc677/foreign_mediator.test.js
@@ -115,9 +115,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        owner,
-        [],
-        [ZERO, ZERO]
+        owner
       ).should.be.rejected
 
       // dailyLimit > maxPerTx
@@ -127,9 +125,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [maxPerTx, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        owner,
-        [],
-        [ZERO, ZERO]
+        owner
       ).should.be.rejected
 
       // maxPerTx > minPerTx
@@ -139,9 +135,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, minPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        owner,
-        [],
-        [ZERO, ZERO]
+        owner
       ).should.be.rejected
 
       // executionDailyLimit > executionMaxPerTx
@@ -151,9 +145,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionDailyLimit],
         maxGasPerTx,
-        owner,
-        [],
-        [ZERO, ZERO]
+        owner
       ).should.be.rejected
 
       // maxGasPerTx > bridge maxGasPerTx
@@ -163,9 +155,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         twoEthers,
-        owner,
-        [],
-        [ZERO, ZERO]
+        owner
       ).should.be.rejected
 
       // not valid owner
@@ -175,9 +165,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        ZERO_ADDRESS,
-        [],
-        [ZERO, ZERO]
+        ZERO_ADDRESS
       ).should.be.rejected
 
       const { logs } = await contract.initialize(
@@ -186,9 +174,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        owner,
-        [],
-        [ZERO, ZERO]
+        owner
       ).should.be.fulfilled
 
       // already initialized
@@ -198,9 +184,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        owner,
-        [],
-        [ZERO, ZERO]
+        owner
       ).should.be.rejected
 
       // Then
@@ -243,9 +227,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        owner,
-        [user2],
-        [ether('0.1'), ZERO]
+        owner
       ).should.be.fulfilled
     })
 
@@ -298,9 +280,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        owner,
-        [],
-        [ZERO, ZERO]
+        owner
       ).should.be.fulfilled
 
       const initialEvents = await getEvents(ambBridgeContract, { event: 'MockedEvent' })
@@ -620,30 +600,11 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         await token.mint(user, '1').should.be.fulfilled
         await token.transfer(contract.address, '1', { from: user }).should.be.fulfilled
 
-        expect(await contract.dailyLimit(token.address)).to.be.bignumber.equal('3')
-        expect(await contract.maxPerTx(token.address)).to.be.bignumber.equal('2')
+        expect(await contract.dailyLimit(token.address)).to.be.bignumber.equal('10000')
+        expect(await contract.maxPerTx(token.address)).to.be.bignumber.equal('100')
         expect(await contract.minPerTx(token.address)).to.be.bignumber.equal('1')
-        expect(await contract.executionDailyLimit(token.address)).to.be.bignumber.equal('3')
-        expect(await contract.executionMaxPerTx(token.address)).to.be.bignumber.equal('2')
-      })
-
-      it('should initialize fees', async () => {
-        const HOME_TO_FOREIGN_FEE = await contract.HOME_TO_FOREIGN_FEE()
-        const FOREIGN_TO_HOME_FEE = await contract.FOREIGN_TO_HOME_FEE()
-        await contract.setFee(HOME_TO_FOREIGN_FEE, ZERO_ADDRESS, ether('0.01'))
-        await contract.setFee(FOREIGN_TO_HOME_FEE, ZERO_ADDRESS, ether('0.02'))
-
-        expect(await contract.getFee(HOME_TO_FOREIGN_FEE, ZERO_ADDRESS)).to.be.bignumber.equal(ether('0.01'))
-        expect(await contract.getFee(FOREIGN_TO_HOME_FEE, ZERO_ADDRESS)).to.be.bignumber.equal(ether('0.02'))
-        expect(await contract.getFee(HOME_TO_FOREIGN_FEE, token.address)).to.be.bignumber.equal(ZERO)
-        expect(await contract.getFee(FOREIGN_TO_HOME_FEE, token.address)).to.be.bignumber.equal(ZERO)
-
-        await token.transfer(contract.address, value, { from: user }).should.be.fulfilled
-
-        expect(await contract.getFee(HOME_TO_FOREIGN_FEE, ZERO_ADDRESS)).to.be.bignumber.equal(ether('0.01'))
-        expect(await contract.getFee(FOREIGN_TO_HOME_FEE, ZERO_ADDRESS)).to.be.bignumber.equal(ether('0.02'))
-        expect(await contract.getFee(HOME_TO_FOREIGN_FEE, token.address)).to.be.bignumber.equal(ether('0.01'))
-        expect(await contract.getFee(FOREIGN_TO_HOME_FEE, token.address)).to.be.bignumber.equal(ether('0.02'))
+        expect(await contract.executionDailyLimit(token.address)).to.be.bignumber.equal('10000')
+        expect(await contract.executionMaxPerTx(token.address)).to.be.bignumber.equal('100')
       })
     })
 
@@ -691,6 +652,24 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         expect(event[0].returnValues.recipient).to.be.equal(user)
         expect(event[0].returnValues.value).to.be.equal(value.toString())
         expect(event[0].returnValues.messageId).to.be.equal(exampleMessageId)
+      })
+
+      it('should not allow to use unregistered tokens', async () => {
+        const otherToken = await ERC20Mock.new('Test', 'TST', 18)
+        await otherToken.mint(contract.address, value)
+        const data = await contract.contract.methods
+          .handleBridgedTokens(otherToken.address, user, value.toString())
+          .encodeABI()
+
+        await ambBridgeContract.executeMessageCall(
+          contract.address,
+          otherSideMediator.address,
+          data,
+          failedMessageId,
+          1000000
+        ).should.be.fulfilled
+
+        expect(await ambBridgeContract.messageCallStatus(failedMessageId)).to.be.equal(false)
       })
     })
 
@@ -896,9 +875,7 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
         [dailyLimit, maxPerTx, minPerTx],
         [executionDailyLimit, executionMaxPerTx],
         maxGasPerTx,
-        owner,
-        [owner],
-        [ZERO, ZERO]
+        owner
       ).should.be.fulfilled
 
       expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(ZERO)
@@ -911,7 +888,6 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
       expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(twoEthers)
 
       await token.transfer(contract.address, halfEther, { from: user }).should.be.fulfilled
-      await contract.setFee(await contract.FOREIGN_TO_HOME_FEE(), token.address, ether('0.1')).should.be.fulfilled
       await contract.setDailyLimit(token.address, ether('5')).should.be.fulfilled
       await contract.setMaxPerTx(token.address, ether('2')).should.be.fulfilled
 
@@ -939,7 +915,6 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
       expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(twoEthers)
 
       await token.transfer(contract.address, halfEther, { from: user }).should.be.fulfilled
-      await contract.setFee(await contract.FOREIGN_TO_HOME_FEE(), token.address, ether('0.1')).should.be.fulfilled
 
       expect(await contract.mediatorBalance(token.address)).to.be.bignumber.equal(halfEther)
       expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(twoEthers.add(halfEther))
@@ -967,301 +942,6 @@ contract('ForeignMultiAMBErc20ToErc677', async accounts => {
 
       events = await getEvents(ambBridgeContract, { event: 'MockedEvent' })
       expect(events.length).to.be.equal(3)
-    })
-  })
-
-  describe('fees management', () => {
-    beforeEach(async () => {
-      await contract.initialize(
-        ambBridgeContract.address,
-        otherSideMediator.address,
-        [dailyLimit, maxPerTx, minPerTx],
-        [executionDailyLimit, executionMaxPerTx],
-        maxGasPerTx,
-        owner,
-        [owner],
-        [ether('0.02'), ether('0.01')]
-      ).should.be.fulfilled
-
-      const initialEvents = await getEvents(ambBridgeContract, { event: 'MockedEvent' })
-      expect(initialEvents.length).to.be.equal(0)
-    })
-
-    it('change reward addresses', async () => {
-      await contract.addRewardAddress(accounts[8], { from: user }).should.be.rejected
-      await contract.addRewardAddress(owner).should.be.rejected
-      await contract.addRewardAddress(accounts[8]).should.be.fulfilled
-
-      expect(await contract.rewardAddressList()).to.be.eql([accounts[8], owner])
-      expect(await contract.rewardAddressCount()).to.be.bignumber.equal('2')
-      expect(await contract.isRewardAddress(owner)).to.be.equal(true)
-      expect(await contract.isRewardAddress(accounts[8])).to.be.equal(true)
-
-      await contract.addRewardAddress(accounts[9]).should.be.fulfilled
-      expect(await contract.rewardAddressList()).to.be.eql([accounts[9], accounts[8], owner])
-      expect(await contract.rewardAddressCount()).to.be.bignumber.equal('3')
-
-      await contract.removeRewardAddress(owner, { from: user }).should.be.rejected
-      await contract.removeRewardAddress(accounts[7]).should.be.rejected
-      await contract.removeRewardAddress(accounts[8]).should.be.fulfilled
-      await contract.removeRewardAddress(accounts[8]).should.be.rejected
-
-      expect(await contract.rewardAddressList()).to.be.eql([accounts[9], owner])
-      expect(await contract.rewardAddressCount()).to.be.bignumber.equal('2')
-      expect(await contract.isRewardAddress(accounts[8])).to.be.equal(false)
-
-      await contract.removeRewardAddress(owner).should.be.fulfilled
-      expect(await contract.rewardAddressList()).to.be.eql([accounts[9]])
-      expect(await contract.rewardAddressCount()).to.be.bignumber.equal('1')
-      expect(await contract.isRewardAddress(owner)).to.be.equal(false)
-
-      await contract.removeRewardAddress(accounts[9]).should.be.fulfilled
-      expect(await contract.rewardAddressList()).to.be.eql([])
-      expect(await contract.rewardAddressCount()).to.be.bignumber.equal('0')
-      expect(await contract.isRewardAddress(accounts[9])).to.be.equal(false)
-    })
-
-    describe('update fee parameters', () => {
-      it('should update default fee value', async () => {
-        const feeType = await contract.HOME_TO_FOREIGN_FEE()
-        await contract.setFee(feeType, ZERO_ADDRESS, ether('0.1'), { from: user }).should.be.rejected
-        await contract.setFee(feeType, ZERO_ADDRESS, ether('1.1'), { from: owner }).should.be.rejected
-        const { logs } = await contract.setFee(feeType, ZERO_ADDRESS, ether('0.1'), { from: owner }).should.be.fulfilled
-
-        expectEventInLogs(logs, 'FeeUpdated')
-        expect(await contract.getFee(feeType, ZERO_ADDRESS)).to.be.bignumber.equal(ether('0.1'))
-        expect(await contract.getFee(await contract.FOREIGN_TO_HOME_FEE(), ZERO_ADDRESS)).to.be.bignumber.equal(
-          ether('0.01')
-        )
-      })
-
-      it('should update default opposite direction fee value', async () => {
-        const feeType = await contract.FOREIGN_TO_HOME_FEE()
-        await contract.setFee(feeType, ZERO_ADDRESS, ether('0.1'), { from: user }).should.be.rejected
-        await contract.setFee(feeType, ZERO_ADDRESS, ether('1.1'), { from: owner }).should.be.rejected
-        const { logs } = await contract.setFee(feeType, ZERO_ADDRESS, ether('0.1'), { from: owner }).should.be.fulfilled
-
-        expectEventInLogs(logs, 'FeeUpdated')
-        expect(await contract.getFee(feeType, ZERO_ADDRESS)).to.be.bignumber.equal(ether('0.1'))
-        expect(await contract.getFee(await contract.HOME_TO_FOREIGN_FEE(), ZERO_ADDRESS)).to.be.bignumber.equal(
-          ether('0.02')
-        )
-      })
-
-      it('should update fee value for registered token', async () => {
-        const feeType = await contract.HOME_TO_FOREIGN_FEE()
-        await token.mint(user, twoEthers, { from: owner }).should.be.fulfilled
-
-        await contract.setFee(feeType, token.address, ether('0.1'), { from: user }).should.be.rejected
-        await contract.setFee(feeType, token.address, ether('1.1'), { from: owner }).should.be.rejected
-        await contract.setFee(feeType, token.address, ether('0.1'), { from: owner }).should.be.rejected
-        await token.transfer(contract.address, value, { from: user }).should.be.fulfilled
-        await contract.setFee(feeType, token.address, ether('0.1'), { from: user }).should.be.rejected
-        await contract.setFee(feeType, token.address, ether('1.1'), { from: owner }).should.be.rejected
-        const { logs } = await contract.setFee(feeType, token.address, ether('0.1'), { from: owner }).should.be
-          .fulfilled
-
-        expectEventInLogs(logs, 'FeeUpdated')
-        expect(await contract.getFee(feeType, token.address)).to.be.bignumber.equal(ether('0.1'))
-        expect(await contract.getFee(await contract.FOREIGN_TO_HOME_FEE(), token.address)).to.be.bignumber.equal(
-          ether('0.01')
-        )
-      })
-
-      it('should update opposite direction fee value for registered token', async () => {
-        const feeType = await contract.FOREIGN_TO_HOME_FEE()
-        await token.mint(user, twoEthers, { from: owner }).should.be.fulfilled
-
-        await contract.setFee(feeType, token.address, ether('0.1'), { from: user }).should.be.rejected
-        await contract.setFee(feeType, token.address, ether('1.1'), { from: owner }).should.be.rejected
-        await contract.setFee(feeType, token.address, ether('0.1'), { from: owner }).should.be.rejected
-        await token.transfer(contract.address, value, { from: user }).should.be.fulfilled
-        await contract.setFee(feeType, token.address, ether('0.1'), { from: user }).should.be.rejected
-        await contract.setFee(feeType, token.address, ether('1.1'), { from: owner }).should.be.rejected
-        const { logs } = await contract.setFee(feeType, token.address, ether('0.1'), { from: owner }).should.be
-          .fulfilled
-
-        expectEventInLogs(logs, 'FeeUpdated')
-        expect(await contract.getFee(feeType, token.address)).to.be.bignumber.equal(ether('0.1'))
-        expect(await contract.getFee(await contract.HOME_TO_FOREIGN_FEE(), token.address)).to.be.bignumber.equal(
-          ether('0.02')
-        )
-      })
-    })
-
-    describe('distribute fee for home => foreign direction', async () => {
-      beforeEach(async () => {
-        await token.mint(user, twoEthers, { from: owner }).should.be.fulfilled
-        expect(await token.balanceOf(user)).to.be.bignumber.equal(twoEthers)
-        await contract.setFee(await contract.FOREIGN_TO_HOME_FEE(), ZERO_ADDRESS, ZERO).should.be.fulfilled
-        await token.transfer(contract.address, value, { from: user }).should.be.fulfilled
-        expect(await token.balanceOf(user)).to.be.bignumber.equal(oneEther)
-      })
-
-      it('should collect and distribute 0% fee', async () => {
-        await contract.setFee(await contract.HOME_TO_FOREIGN_FEE(), token.address, ZERO).should.be.fulfilled
-
-        const data = await contract.contract.methods
-          .handleBridgedTokens(token.address, user, value.toString())
-          .encodeABI()
-
-        await ambBridgeContract.executeMessageCall(
-          contract.address,
-          otherSideMediator.address,
-          data,
-          exampleMessageId,
-          1000000
-        ).should.be.fulfilled
-
-        expect(await ambBridgeContract.messageCallStatus(exampleMessageId)).to.be.equal(true)
-        expect(await contract.totalExecutedPerDay(token.address, currentDay)).to.be.bignumber.equal(value)
-
-        const event = await getEvents(contract, { event: 'TokensBridged' })
-        expect(event.length).to.be.equal(1)
-        expect(event[0].returnValues.token).to.be.equal(token.address)
-        expect(event[0].returnValues.recipient).to.be.equal(user)
-        expect(event[0].returnValues.value).to.be.equal(value.toString())
-        expect(event[0].returnValues.messageId).to.be.equal(exampleMessageId)
-
-        const feeEvents = await getEvents(contract, { event: 'FeeDistributed' })
-        expect(feeEvents.length).to.be.equal(0)
-
-        expect(await token.balanceOf(user)).to.be.bignumber.equal(twoEthers)
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(ZERO)
-        expect(await token.balanceOf(owner)).to.be.bignumber.equal(ZERO)
-      })
-
-      it('should collect and distribute 2% fee', async () => {
-        const data = await contract.contract.methods
-          .handleBridgedTokens(token.address, user, value.toString())
-          .encodeABI()
-
-        await ambBridgeContract.executeMessageCall(
-          contract.address,
-          otherSideMediator.address,
-          data,
-          exampleMessageId,
-          1000000
-        ).should.be.fulfilled
-
-        expect(await ambBridgeContract.messageCallStatus(exampleMessageId)).to.be.equal(true)
-        expect(await contract.totalExecutedPerDay(token.address, currentDay)).to.be.bignumber.equal(value)
-
-        const event = await getEvents(contract, { event: 'TokensBridged' })
-        expect(event.length).to.be.equal(1)
-        expect(event[0].returnValues.token).to.be.equal(token.address)
-        expect(event[0].returnValues.recipient).to.be.equal(user)
-        expect(event[0].returnValues.value).to.be.equal(ether('0.98').toString())
-        expect(event[0].returnValues.messageId).to.be.equal(exampleMessageId)
-
-        const feeEvents = await getEvents(contract, { event: 'FeeDistributed' })
-        expect(feeEvents.length).to.be.equal(1)
-
-        expect(await token.balanceOf(user)).to.be.bignumber.equal(ether('1.98'))
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(ZERO)
-        expect(await token.balanceOf(owner)).to.be.bignumber.equal(ether('0.02'))
-      })
-
-      it('should collect and distribute 2% fee between two reward addresses', async () => {
-        await contract.addRewardAddress(accounts[9]).should.be.fulfilled
-        expect(await contract.rewardAddressCount()).to.be.bignumber.equal('2')
-
-        const data = await contract.contract.methods
-          .handleBridgedTokens(token.address, user, ether('0.100000000000000050').toString(10))
-          .encodeABI()
-
-        await ambBridgeContract.executeMessageCall(
-          contract.address,
-          otherSideMediator.address,
-          data,
-          exampleMessageId,
-          1000000
-        ).should.be.fulfilled
-
-        expect(await ambBridgeContract.messageCallStatus(exampleMessageId)).to.be.equal(true)
-        expect(await contract.totalExecutedPerDay(token.address, currentDay)).to.be.bignumber.equal(
-          ether('0.100000000000000050')
-        )
-
-        const event = await getEvents(contract, { event: 'TokensBridged' })
-        expect(event.length).to.be.equal(1)
-
-        const feeEvents = await getEvents(contract, { event: 'FeeDistributed' })
-        expect(feeEvents.length).to.be.equal(1)
-
-        expect(await token.balanceOf(user)).to.be.bignumber.equal(ether('1.098000000000000049'))
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(ether('0.899999999999999950'))
-        const balance1 = (await token.balanceOf(owner)).toString()
-        const balance2 = (await token.balanceOf(accounts[9])).toString()
-        expect(
-          (balance1 === '1000000000000001' && balance2 === '1000000000000000') ||
-            (balance1 === '1000000000000000' && balance2 === '1000000000000001')
-        ).to.be.equal(true)
-      })
-    })
-
-    describe('distribute fee for foreign => home direction', async () => {
-      beforeEach(async () => {
-        await token.mint(user, twoEthers).should.be.fulfilled
-      })
-
-      it('should collect and distribute 0% fee', async () => {
-        await contract.setFee(await contract.FOREIGN_TO_HOME_FEE(), ZERO_ADDRESS, ZERO).should.be.fulfilled
-
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(ZERO)
-        await token.transfer(contract.address, value, { from: user })
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(value)
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(value)
-        await token.transfer(contract.address, value, { from: user })
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(twoEthers)
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(twoEthers)
-
-        const feeEvents = await getEvents(contract, { event: 'FeeDistributed' })
-        expect(feeEvents.length).to.be.equal(0)
-      })
-
-      it('should collect and distribute 1% fee', async () => {
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(ZERO)
-        await token.transfer(contract.address, value, { from: user })
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(value)
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(ether('0.99'))
-        expect(await token.balanceOf(owner)).to.be.bignumber.equal(ether('0.01'))
-        await token.transfer(contract.address, value, { from: user })
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(twoEthers)
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(ether('1.98'))
-        expect(await token.balanceOf(owner)).to.be.bignumber.equal(ether('0.02'))
-
-        const feeEvents = await getEvents(contract, { event: 'FeeDistributed' })
-        expect(feeEvents.length).to.be.equal(2)
-      })
-
-      it('should collect and distribute 1% fee between two reward addresses', async () => {
-        await contract.addRewardAddress(accounts[9]).should.be.fulfilled
-        expect(await contract.rewardAddressCount()).to.be.bignumber.equal('2')
-
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(ZERO)
-        await token.transfer(contract.address, ether('0.200000000000000100'), { from: user })
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(
-          ether('0.200000000000000100')
-        )
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(ether('0.198000000000000099'))
-
-        const balance1 = (await token.balanceOf(owner)).toString()
-        const balance2 = (await token.balanceOf(accounts[9])).toString()
-        expect(
-          (balance1 === '1000000000000001' && balance2 === '1000000000000000') ||
-            (balance1 === '1000000000000000' && balance2 === '1000000000000001')
-        ).to.be.equal(true)
-
-        await token.transfer(contract.address, value, { from: user }).should.be.fulfilled
-        expect(await contract.totalSpentPerDay(token.address, currentDay)).to.be.bignumber.equal(
-          ether('1.200000000000000100')
-        )
-        expect(await token.balanceOf(contract.address)).to.be.bignumber.equal(ether('1.188000000000000099'))
-
-        const feeEvents = await getEvents(contract, { event: 'FeeDistributed' })
-        expect(feeEvents.length).to.be.equal(2)
-      })
     })
   })
 })
